Move mint/ repay vai to Dashboard page dir
