+++ src/pages/Dashboard/MintRepayVai/Mint/styles.ts
